refactor: rename a file for consistency
